Modify test_utils.test_preprocessing API to return entire dataset, and fix a bug.
The current behavior for test_utils.test_processing is to return a single item
from a dataset (the first). This behavior prevents a more general use of this
utility when we'd like to create tests that require multiple dataset items.

Instead of creating an alternative function, or flag-guard the return of this
utility, I decided to keep things simple and unconditionally return the entire
dataset as a list. Callers were updated to simply take the first item of the
returned value.

In addition to API change, this CL:

1) Fixes a bug where get_dataset() did not specify a split, and was thus
  _always_ set to the default split (TRAIN) irrespective of the split set in
  the raw_data.

  We now use the split specified in the raw_data map to get_dataset().

2) Do not allow more than one split in raw_data for now. Its unclear what the
   correct behavior is (concatenating the datasets? which order?). We can
   revisit this if there should be an user for this behavior.

3) Prevent data shuffling which is particularly tricky in tests.

PiperOrigin-RevId: 377327225